Convert 'year' field to integer in FinalResultsImport.
Previously, the 'year' field was treated as a string, which could lead to inconsistencies in data handling. This change ensures it is cast to an integer, improving type consistency and reducing potential errors.

+++ app/Imports/FinalResultsImport.php
@@ -128,7 +128,7 @@ final class FinalResultsImport implements ToModel, WithBatchInserts, WithCalcula
             'session' => Str::trim($row[$this->headings['session']]),
             'sn' => (int) Str::trim($row[$this->headings['sn']]),
             'total' => (int) Str::trim($row[$this->headings['total']]),
-            'year' => Str::trim($row[$this->headings['year']]),
+            'year' => (int) Str::trim($row[$this->headings['year']]),
         ];
     }
 }
